OMS-887 evengsdk renamed to py-eveng
