+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include README.md
 
-include src/evengsdk/schemas/lab-schema.json
+include src/py_eveng/schemas/lab-schema.json
