+++ Makefile
@@ -59,7 +59,7 @@ test-all: ## run tests on every Python version with tox
 	tox
 
 coverage: ## check code coverage quickly with the default Python
-	coverage run --source evengsdk -m pytest
+	coverage run --source py_eveng -m pytest
 	coverage report -m
 	coverage html
 	$(BROWSER) htmlcov/index.html
