+++ pytest.ini
@@ -4,7 +4,7 @@ addopts =
     --strict-config
     -ra
     -m "not xfail"
-    --cov=evengsdk
+    --cov=py-eveng
     --cov-report html
 
 testpaths = src/tests
