--- src/evengsdk/__init__.py
@@ -1,1 +1,0 @@
-from evengsdk.cli.version import __version__  # noqa: F401
