+++ src/py_eveng/__init__.py
@@ -1,0 +1,1 @@
+from py_eveng.cli.version import __version__  # noqa: F401
