+++ src/py_eveng/cli/users/commands.py
@@ -1,10 +1,10 @@
 import click
 from pygments.lexer import default
 
-from evengsdk.cli.common import list_sub_command
-from evengsdk.cli.console import cli_print_output, console
-from evengsdk.cli.utils import get_client
-from evengsdk.exceptions import EvengApiError, EvengHTTPError
+from py_eveng.cli.common import list_sub_command
+from py_eveng.cli.console import cli_print_output, console
+from py_eveng.cli.utils import get_client
+from py_eveng.exceptions import EvengApiError, EvengHTTPError
 
 
 def user_sub_command(subcommand):
