+++ src/py_eveng/exceptions.py
