+++ src/py_eveng/templates/inventory.ini.j2
