+++ src/tests/api/test_api_network.py
@@ -1,6 +1,6 @@
 import pytest
 
-from evengsdk.exceptions import EvengHTTPError
+from py_eveng.exceptions import EvengHTTPError
 
 
 @pytest.mark.usefixtures("setup_lab")
