+++ src/tests/cli/test_cli_lab.py
@@ -4,7 +4,7 @@ import pytest
 import yaml
 from click.testing import CliRunner, Result
 
-from evengsdk.cli.cli import main as cli
+from py_eveng.cli.cli import main as cli
 
 
 @pytest.fixture
